fix: clean and format

+++ contracts/DXswapFactory.sol
@@ -1,8 +1,8 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 pragma solidity ^0.8.0;
 
-import './interfaces/IDXswapFactory.sol';
-import './DXswapPair.sol';
+import "./interfaces/IDXswapFactory.sol";
+import "./DXswapPair.sol";
 
 contract DXswapFactory is IDXswapFactory {
     address public feeTo;
@@ -13,21 +13,24 @@ contract DXswapFactory is IDXswapFactory {
     mapping(address => mapping(address => address)) public getPair;
     address[] public allPairs;
 
+    /*
+    * @dev Came from parent contract
+    */
     // event PairCreated(address indexed token0, address indexed token1, address pair, uint);
 
     constructor(address _feeToSetter) {
         feeToSetter = _feeToSetter;
     }
 
-    function allPairsLength() external view override returns (uint) {
+    function allPairsLength() external view override returns (uint256) {
         return allPairs.length;
     }
 
     function createPair(address tokenA, address tokenB) external override returns (address pair) {
-        require(tokenA != tokenB, 'DXswapFactory: IDENTICAL_ADDRESSES');
+        require(tokenA != tokenB, "DXswapFactory: IDENTICAL_ADDRESSES");
         (address token0, address token1) = tokenA < tokenB ? (tokenA, tokenB) : (tokenB, tokenA);
-        require(token0 != address(0), 'DXswapFactory: ZERO_ADDRESS');
-        require(getPair[token0][token1] == address(0), 'DXswapFactory: PAIR_EXISTS'); // single check is sufficient
+        require(token0 != address(0), "DXswapFactory: ZERO_ADDRESS");
+        require(getPair[token0][token1] == address(0), "DXswapFactory: PAIR_EXISTS"); // single check is sufficient
         bytes memory bytecode = type(DXswapPair).creationCode;
         bytes32 salt = keccak256(abi.encodePacked(token0, token1));
         assembly {
@@ -41,23 +44,23 @@ contract DXswapFactory is IDXswapFactory {
     }
 
     function setFeeTo(address _feeTo) external override {
-        require(msg.sender == feeToSetter, 'DXswapFactory: FORBIDDEN');
+        require(msg.sender == feeToSetter, "DXswapFactory: FORBIDDEN");
         feeTo = _feeTo;
     }
 
     function setFeeToSetter(address _feeToSetter) external override {
-        require(msg.sender == feeToSetter, 'DXswapFactory: FORBIDDEN');
+        require(msg.sender == feeToSetter, "DXswapFactory: FORBIDDEN");
         feeToSetter = _feeToSetter;
     }
 
     function setProtocolFee(uint8 _protocolFeeDenominator) external override {
-        require(msg.sender == feeToSetter, 'DXswapFactory: FORBIDDEN');
-        require(_protocolFeeDenominator > 0, 'DXswapFactory: FORBIDDEN_FEE');
+        require(msg.sender == feeToSetter, "DXswapFactory: FORBIDDEN");
+        require(_protocolFeeDenominator > 0, "DXswapFactory: FORBIDDEN_FEE");
         protocolFeeDenominator = _protocolFeeDenominator;
     }
 
     function setSwapFee(address _pair, uint32 _swapFee) external override {
-        require(msg.sender == feeToSetter, 'DXswapFactory: FORBIDDEN');
+        require(msg.sender == feeToSetter, "DXswapFactory: FORBIDDEN");
         IDXswapPair(_pair).setSwapFee(_swapFee);
     }
 }
